requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
beautifulsoup4==4.9.1
black==19.10b0
boto3==1.14.34
botocore==1.17.34
click==7.1.2
dj-database-url==0.5.0
Django==3.0.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.1.1
django-crispy-forms==1.9.1
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
importlib-metadata==1.6.1
jmespath==0.10.0
pathspec==0.8.0
Pillow==7.1.2
python-dateutil==2.8.1
pytz==2020.1
regex==2020.6.8
s3transfer==0.3.3
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
whitenoise==5.1.0
zipp==3.1.0
